--- src/data/calculator.js
@@ -1,5 +1,0 @@
-const calculator = () => {
-    return 0;
-}
-
-export default calculator
