+++ src/data/useGetProduct.ts
@@ -1,10 +1,9 @@
 import { useGetProducts } from "../context/productContext";
-
-const useGetProduct = (id) => {
+const useGetProduct = (id: string) => {
   const products = useGetProducts();
   if (!id) throw new Error("missing ID in useGetProduct call");
 
-  const product = products.item.find((product) => product.id === id);
+  const product = products?.find((product) => product.id === id);
   return { item: product };
 };
