+++ src/interfaces.ts
@@ -1,0 +1,10 @@
+export interface Product {
+  id?: string;
+  name?: string;
+  description?: string;
+  details?: string;
+  price?: number;
+  vat?: number;
+  images?: string[];
+  category?: string;
+}
